+++ lib/src/analyzers/unused_files_analyzer/reporters/reporters_list/json/unused_files_json_reporter.dart
@@ -5,19 +5,21 @@ import 'package:meta/meta.dart';
 
 import '../../../../../reporters/models/json_reporter.dart';
 import '../../../models/unused_files_file_report.dart';
+import '../../report_params.dart';
 
 /// Unused files JSON reporter.
 ///
 /// Use it to create reports in JSON format.
 @immutable
 class UnusedFilesJsonReporter
-    extends JsonReporter<UnusedFilesFileReport, void> {
+    extends JsonReporter<UnusedFilesFileReport, void, ReportParams> {
   const UnusedFilesJsonReporter(IOSink output) : super(output, 2);
 
   @override
   Future<void> report(
     Iterable<UnusedFilesFileReport> records, {
     Iterable<void> summary = const [],
+    ReportParams? additionalParams,
   }) async {
     if (records.isEmpty) {
       return;
@@ -27,6 +29,7 @@ class UnusedFilesJsonReporter
       'formatVersion': formatVersion,
       'timestamp': getTimestamp(),
       'unusedFiles': records.map(_analysisRecordToJson).toList(),
+      'automaticallyDeleted': additionalParams?.deleteUnusedFiles ?? false,
     });
 
     output.write(encodedReport);